jsonrpc.js を class 構文で書き直した．
