+++ nkjmlab-go-webapp/pom.xml
@@ -2,7 +2,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.nkjmlab</groupId>
   <artifactId>nkjmlab-go-webapp</artifactId>
-  <version>0.9.5</version>
+  <version>0.9.6</version>
   <description>Go Web Application</description>
   <url>https://github.com/nkjmlab/nkjmlab-go-webapp</url>
   <licenses>
@@ -26,7 +26,7 @@
   </scm>
 
   <properties>
-    <nkjmlab-utils-version>0.8.3</nkjmlab-utils-version>
+    <nkjmlab-utils-version>0.8.4</nkjmlab-utils-version>
     <thymeleaf-version>3.0.15.RELEASE</thymeleaf-version>
     <sorm4j-version>1.4.9</sorm4j-version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -316,8 +316,17 @@
       </exclusions>
     </dependency>
 
-
-
+    <dependency>
+      <groupId>org.webjars.npm</groupId>
+      <artifactId>ua-parser-js</artifactId>
+      <version>1.0.2</version>
+      <exclusions>
+        <exclusion>
+          <groupId>*</groupId>
+          <artifactId>*</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
 
     <!-- https://mvnrepository.com/artifact/org.junit.jupiter/junit-jupiter-engine -->
     <dependency>
