+++ nkjmlab-go-webapp/src/main/resources/templates/fragment/head.html
@@ -36,9 +36,8 @@
 <script src="/webjars/firebase/9.9.0/firebase-firestore-compat.js"></script>
 
 
-<script src="../js/lib/parseUri.js"></script>
-<script src="../js/lib/ua-parser.min.js"></script>
-<script th:src="|../js/lib/jsonrpc.js?${js_lib_jsonrpc_js}|"></script>
+<script src="/webjars/ua-parser-js/1.0.2/dist/ua-parser.min.js"></script>
+<script th:src="|../js/lib/jsonrpc-client.js?${js_lib_jsonrpc_client_js}|"></script>
 <script th:src="|../js/utils.js?${js_utils_js}|"></script>
 <script th:src="|../js/firebase.js?${js_firebase_js}|"></script>
